Stop convering stdout to a string, send Buffer to window ...
Fix performance issues with string copying:

Previously the main process sent stdout from the shell as a string to the window, the renderer would then store it and each render concat the string with the current users input. When the string grew to a reasoanble size this copying would make the app unusable.
Instead we send the buffer from stdout without converting it to a string, and display each chunk of data in it's own <pre> container.

This stops the string being copied (potentially a very large string!) every time a key is pressed.